requirements.txt
alabaster==0.7.12
anyio==3.3.2
argon2-cffi==21.1.0
astroid==2.6.6
atomicwrites==1.4.0
attrs==20.3.0
auxlib==0.0.43
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
boto==2.49.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
colorama==0.4.4
conda==4.2.7
coverage==5.5
cpplint==1.5.5
cycler==0.10.0
Cython==0.29.24
cytools==1.0.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
docutils==0.16
entrypoints==0.3
ffmpeg==1.4
ffmpeg-python==0.2.0
flameprof==0.4
flatbuffers==2.0
fonttools==4.28.1
future==0.18.2
gitdb==4.0.7
GitPython==3.1.24
greenlet==1.1.1
idna==3.2
imagesize==1.2.0
importlib-metadata==4.8.1
importlib-resources==5.2.2
iniconfig==1.1.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.9.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-book==0.11.3
jupyter-cache==0.4.3
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyter-server-mathjax==0.2.3
jupyter-sphinx==0.3.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
jupytext==1.10.3
kiwisolver==1.3.2
latexcodec==2.0.1
lazy-object-proxy==1.6.0
linkify-it-py==1.0.1
markdown-it-py==0.6.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mdit-py-plugins==0.2.6
mistune==0.8.4
mpmath==1.2.1
myst-nb==0.12.3
myst-parser==0.13.7
nbclient==0.5.4
nbconvert==5.6.1
nbdime==3.1.0
nbformat==5.1.3
nbstripout==0.5.0
nbval==0.9.6
nest-asyncio==1.5.1
ninja==1.10.2.2
notebook==6.4.4
numpy==1.21.2
onnx==1.10.1
onnx-simplifier==0.3.6
onnxoptimizer==0.2.6
onnxruntime==1.9.0
opencv-python==4.5.3.56
openpifpaf==0.12.14
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pluggy==1.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
py==1.10.0
pybtex==0.24.0
pybtex-docutils==1.0.1
pycodestyle==2.7.0
pycparser==2.20
pydata-sphinx-theme==0.6.3
Pygments==2.10.0
pylint==2.9.3
pyparsing==2.4.7
pyrsistent==0.18.0
pysparkling==0.6.1
pytest==6.2.5
python-dateutil==2.8.2
python-json-logger==2.0.2
pytoolz==0.1.7
pytz==2021.1
pywin32==302
pywinpty==1.1.6
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
requests-unixsocket==0.2.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.5.4
sphinx-book-theme==0.1.5
sphinx-comments==0.0.3
sphinx-copybutton==0.4.0
sphinx-external-toc==0.2.3
sphinx-jupyterbook-latex==0.4.2
sphinx-multitoc-numbering==0.1.3
sphinx-panels==0.5.2
sphinx-thebe==0.0.10
sphinx-togglebutton==0.2.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.2.1
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.25
sympy==1.9
terminado==0.12.1
testpath==0.5.0
thop==0.0.31.post2005241907
timm==0.4.12
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
torch==1.9.1
torchaudio==0.10.0+cu102
torchvision==0.10.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
uc-micro-py==1.0.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.5.0
